pyproject.toml
[tool.poetry]
name = "stacker"
version = "0.1.0"
description = "configuration and dockerfile generation for stacks"
authors = ["unrahul <rahul.unnikrishnan.nair@intel.com>"]
license = "Apache-2.0"
packages = [
  { include = "stacker", from = "."}
]
[tool.poetry.dependencies]
python = "^3.6"
jinja2 = "^2.11.1"
"ruamel.yaml" = "^0.16.7"
fire = "^0.2.1"

[tool.poetry.scripts]
stacker = "stacker:main"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
